run.py
"""Entry point local: python run.py"""
import os
from pathlib import Path

from dotenv import load_dotenv
from app import create_app

BASE_DIR = Path(__file__).resolve().parent
load_dotenv(BASE_DIR / ".env")

app = create_app()

if __name__ == "__main__":
    app.run(host="0.0.0.0",
            port=int(os.getenv("PORT", "5000")),
            debug=os.getenv("FLASK_DEBUG", "false").lower() == "true")
