requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Cors==4.0.1
SQLAlchemy==2.0.34
PyMySQL==1.1.1
cryptography==43.0.1
PyJWT==2.9.0
pyotp==2.9.0
Werkzeug==3.0.4
python-dotenv==1.0.1
gunicorn==22.0.0
requests==2.32.3
openai>=1.30.0
openpyxl>=3.1.2
